bootstrap_admin/requirements.txt
psycopg2-binary==2.9.10

argon2-cffi==25.1.0
